feat: refactor to first API
