+++ example/routes/foo-[id].js
