+++ example/routes/foo/bar.js
