+++ example/routes/no-js.js
