+++ limette/server/app.ts
@@ -1,0 +1,25 @@
+import { Application } from "@oak/oak/application";
+import { staticMiddleware } from "./utils.ts";
+import { router } from "./router.ts";
+
+export class LimetteApp {
+  #app;
+
+  constructor() {
+    this.#app = new Application();
+
+    this.#app.use(staticMiddleware);
+
+    this.#app.use(router.routes());
+    this.#app.use(router.allowedMethods());
+  }
+
+  use(...args) {
+    this.#app.use(...args);
+    return this;
+  }
+
+  listen(...args) {
+    return this.#app.listen(...args);
+  }
+}
