+++ limette/server/rendering/patch-custom-elements.ts
@@ -1,3 +1,5 @@
+// CURRENTLY NOT USED
+
 const customElementsDefine = customElements.define;
 
 /**
